feat: add core data for persistence data

+++ Jet2TT/Resource/Extension/Int.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-extension Int {
+extension Int64 {
     
     func formatCounts() -> String {
         let number = Double(self)
